:recycle: Move SuccessResponse and ErrorResponse types to common file
