+++ app/auth/lib/actions/register.ts
@@ -2,7 +2,7 @@
 
 import { RegisterDto } from '@/auth/lib/types/register.types';
 import { User } from '@/auth/lib/types/user.types';
-import { APIResponse } from '@/lib/types/api-response.types';
+import { APIResponse } from '@/lib/types/api-responses.types';
 
 export async function registerAction(dto: RegisterDto): Promise<APIResponse<User>> {
   const res = await fetch(`${process.env.API_URL}/auth/register`, {
