+++ app/lib/types/api-responses.types.ts
@@ -1,0 +1,18 @@
+export type APIResponse<D = any, M = any> = {
+  data: D;
+  meta: M;
+};
+
+export type SuccessResponse<T> = {
+  isSuccess: true;
+  isError: false;
+  error: null;
+  data: T;
+};
+
+export type ErrorResponse = {
+  isSuccess: false;
+  isError: true;
+  error: Error;
+  data?: undefined;
+};
